Version 1.0 Added: Configuration Editor, which allows to edit configurations in a graphical user interface. Contains GUIs for most commands supported by E-APO, but as lines can also be edited directly, all commands can be used. Added: Command "GraphicEQ", which can act as a regular, fixed-band graphic equalizer (via the GUI) but also supports variable bands so that any desired frequency response can be specified. Internally, it is implemented via convolution with a generated IR. Added: Command "Convolution", which allows to convolve the signal with a user-provided impulse response to achieve e.g. equalization or reverberation effects. Improved: Configurator performs checks for registry values that are needed for the operation of E-APO, which may be changed by driver installations, fixing the values if necessary. Improved: Configurator shows which is the default device as a hint to the user.
git-svn-id: svn://svn.code.sf.net/p/equalizerapo/code/trunk@43 177c5db2-39f6-4273-851a-b0754c4d5937

+++ Benchmark/Benchmark.cpp
@@ -17,6 +17,7 @@
 	51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
 
+#include "stdafx.h"
 #ifdef DEBUG
 #include <stdlib.h>
 #include <crtdbg.h>
@@ -214,7 +215,7 @@ int main(int argc, char** argv)
 
 		printf("\nWriting output to %s\n", output.c_str());
 
-		SF_INFO info = {frameCount, sampleRate, channelCount, SF_FORMAT_WAV | SF_FORMAT_PCM_16, 0};
+		SF_INFO info = {frameCount, (int)sampleRate, (int)channelCount, SF_FORMAT_WAV | SF_FORMAT_PCM_16, 0};
 		SNDFILE* outFile = sf_open(output.c_str(), SFM_WRITE, &info);
 		if(outFile == NULL)
 		{
